+++ test/e2e-runner/src/linux.rs
@@ -65,5 +65,8 @@ pub fn run_linux_e2e() {
         LinuxScenarios::AllOhisNoService(args) => {
             scenarios::all_ohis_no_service::test_all_ohis_no_service(args);
         }
+        LinuxScenarios::Uninstall(args) => {
+            scenarios::uninstall::test_uninstall_script(args);
+        }
     };
 }